+++ changelog.md
@@ -1,3 +1,12 @@
+# 0.0.41
+- kotlin 1.4.10
+- more extensions (arrays & collections)
+- minor code cleanup
+- removed list extensions that were also on collection's
+- naming updated to more "assert" rather than assertEqual (what else should you assert? equals is redundant)
+- removed number.assert as all numbers in kotlin is comparable
+
+
 # 0.0.40
 - kotlin 1.4.0
 - updated contracts